MNT: Clean up ad detector classes
