+++ pcdsdevices/pim.py
@@ -7,39 +7,12 @@ yag position and a camera to view the yag.
 """
 import logging
 
-from ophyd import Component as Cmp, FormattedComponent as FCmp
+from ophyd import FormattedComponent as FCpt
 
-from .areadetector.detectors import PulnixDetector
-from .areadetector.plugins import ImagePlugin, StatsPlugin
+from .areadetector.detectors import DefaultAreaDetector
 from .inout import InOutRecordPositioner
 
 logger = logging.getLogger(__name__)
-
-
-class PIMPulnixDetector(PulnixDetector):
-    """
-    Pulnix detector that is used in the PIM.
-
-    Plugins should be added on an as needed basis here.
-    """
-    image1 = Cmp(ImagePlugin, ":IMAGE1:", read_attrs=['array_data'])
-    image2 = Cmp(ImagePlugin, ":IMAGE2:", read_attrs=['array_data'])
-    stats2 = Cmp(StatsPlugin, ":Stats2:", read_attrs=['centroid',
-                                                      'mean_value'])
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.image1.stage_sigs[self.image1.nd_array_port] = 'CAM'
-        self.image1.stage_sigs[self.image1.blocking_callbacks] = 'Yes'
-        self.image1.stage_sigs[self.image1.enable] = 1
-
-        self.stats2.stage_sigs[self.stats2.nd_array_port] = 'CAM'
-        self.stats2.stage_sigs[self.stats2.blocking_callbacks] = 'Yes'
-        self.stats2.stage_sigs[self.stats2.compute_statistics] = 'Yes'
-        self.stats2.stage_sigs[self.stats2.compute_centroid] = 'Yes'
-        self.stats2.stage_sigs[self.stats2.centroid_threshold] = 200
-        self.stats2.stage_sigs[self.stats2.min_callback_time] = 0.0
-        self.stats2.stage_sigs[self.stats2.enable] = 1
 
 
 class PIMMotor(InOutRecordPositioner):
@@ -78,8 +51,7 @@ class PIM(PIMMotor):
         The EPICS base PV of the detector. If None, it will be inferred from
         the motor prefix
     """
-    detector = FCmp(PIMPulnixDetector, "{self._prefix_det}",
-                    read_attrs=['stats2'])
+    detector = FCpt(DefaultAreaDetector, "{self._prefix_det}")
 
     _default_read_attrs = ['state', 'readback', 'detector']
